finished JS changes for filters more user control over filters changed CSS for filter divs added total_sp to model, bigint for SP total

+++ bazaar/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Character(models.Model):
     name = models.CharField(max_length=64)
     skills = models.ManyToManyField('CharSkill',related_name='learned_by')
-    total_sp = models.IntegerField()
+    total_sp = models.BigIntegerField()
 
 class Skill(models.Model):
     name = models.CharField(max_length=64)
